+++ test/support/testValues.js
@@ -1,7 +1,7 @@
 module.exports = {
-    host         : "192.168.2.113",
+    host         : "192.168.2.158",
     username     : "08a902b95915cdd9b75547cb50892dc4",
-    lightsCount  : 7,
+    lightsCount  : 8,
     locateTimeout: 7000,
     maxScheduleNameLength: 32,
     testLightId: 4